del volume context from remotion root add editor
