+++ packages/cli/src/editor/components/Editor.tsx
@@ -1,6 +1,10 @@
 import {PlayerInternals} from '@remotion/player';
 import React, {useMemo, useState} from 'react';
-import {Internals} from 'remotion';
+import {
+	Internals,
+	MediaVolumeContextValue,
+	SetMediaVolumeContextValue,
+} from 'remotion';
 import styled from 'styled-components';
 import {
 	CheckerboardContext,
@@ -38,6 +42,8 @@ export const Editor: React.FC = () => {
 	const [richTimeline, setRichTimeline] = useState(() =>
 		loadRichTimelineOption()
 	);
+	const [mediaMuted, setMediaMuted] = useState<boolean>(false);
+	const [mediaVolume, setMediaVolume] = useState<number>(1);
 
 	const previewSizeCtx = useMemo(() => {
 		return {
@@ -58,6 +64,20 @@ export const Editor: React.FC = () => {
 		};
 	}, [richTimeline]);
 
+	const mediaVolumeContextValue = useMemo((): MediaVolumeContextValue => {
+		return {
+			mediaMuted,
+			mediaVolume,
+		};
+	}, [mediaMuted, mediaVolume]);
+
+	const setMediaVolumeContextValue = useMemo((): SetMediaVolumeContextValue => {
+		return {
+			setMediaMuted,
+			setMediaVolume,
+		};
+	}, []);
+
 	if (!Root) {
 		throw new Error('Root has not been registered. ');
 	}
@@ -66,27 +86,37 @@ export const Editor: React.FC = () => {
 		<RichTimelineContext.Provider value={richTimelineCtx}>
 			<CheckerboardContext.Provider value={checkerboardCtx}>
 				<PreviewSizeContext.Provider value={previewSizeCtx}>
-					<PlayerInternals.PlayerEventEmitterContext.Provider value={emitter}>
-						<Background>
-							<Root />
-							<UpdateCheck />
-							<SplitterContainer
-								orientation="horizontal"
-								id="top-to-bottom"
-								maxFlex={0.9}
-								minFlex={0.2}
-								defaultFlex={0.75}
+					<Internals.MediaVolumeContext.Provider
+						value={mediaVolumeContextValue}
+					>
+						<Internals.SetMediaVolumeContext.Provider
+							value={setMediaVolumeContextValue}
+						>
+							<PlayerInternals.PlayerEventEmitterContext.Provider
+								value={emitter}
 							>
-								<SplitterElement type="flexer">
-									<TopPanel />
-								</SplitterElement>
-								<SplitterHandle />
-								<SplitterElement type="anti-flexer">
-									<Timeline />
-								</SplitterElement>
-							</SplitterContainer>
-						</Background>
-					</PlayerInternals.PlayerEventEmitterContext.Provider>
+								<Background>
+									<Root />
+									<UpdateCheck />
+									<SplitterContainer
+										orientation="horizontal"
+										id="top-to-bottom"
+										maxFlex={0.9}
+										minFlex={0.2}
+										defaultFlex={0.75}
+									>
+										<SplitterElement type="flexer">
+											<TopPanel />
+										</SplitterElement>
+										<SplitterHandle />
+										<SplitterElement type="anti-flexer">
+											<Timeline />
+										</SplitterElement>
+									</SplitterContainer>
+								</Background>
+							</PlayerInternals.PlayerEventEmitterContext.Provider>
+						</Internals.SetMediaVolumeContext.Provider>
+					</Internals.MediaVolumeContext.Provider>
 				</PreviewSizeContext.Provider>
 			</CheckerboardContext.Provider>
 		</RichTimelineContext.Provider>
